refactor(core): extract inline prop types to types.ts (simple cases)
Checkbox, Dropdown, Modal, Pagination, Radio, SearchBar, SegmentButton,
Tab, TimePicker, Tooltip — move inline IXxxProps into dedicated types.ts.
No runtime change.
