+++ packages/core/src/components/Checkbox/types.ts
@@ -1,0 +1,5 @@
+import type { CheckboxProps as MantineCheckboxProps } from '@mantine/core';
+
+export interface ICheckboxProps extends MantineCheckboxProps {
+  size?: 'sm' | 'md' | 'lg';
+}
