+++ packages/core/src/components/Radio/types.ts
@@ -1,0 +1,5 @@
+import type { RadioProps as MantineRadioProps } from '@mantine/core';
+
+export interface IRadioProps extends MantineRadioProps {
+  size?: 'sm' | 'md' | 'lg';
+}
