+++ packages/core/src/components/Tooltip/types.ts
@@ -1,0 +1,6 @@
+import type { TooltipProps as MantineTooltipProps } from '@mantine/core';
+
+export interface ITooltipProps extends MantineTooltipProps {
+  title?: string;
+  content: string;
+}
